Completely rewritten to separate command construction from execution which enables other commands (e.g., sort) to use and compose the functionality directly from within python.  This uses sh's bake facility to construct commands without actually running them.
Stream-lined and added various new utilities, some of which should
move to a more central location such as cli_entry.py.  Moved all
the functionality outside of `run' and made the `run' signature
more readable and robust.

Improved sh buffer sizing which brings the speed up to be the same
as when an equivalent decomp/comp pipe is run in the shell directly.