Changed line endings created by output writer.
We are forcing csv.write to create \n line endings now (instead of \r\n),
since the former caused us problems when we reimported such data into sqlite.
The \r characters wound up in the values of the last column, the database
worked transparently around that (e.g., in where clause restrictions), but
then the output writer complained that it didn't know how to quote those.
Even though we are now handling \r\n on graph input, we are forcing \n for
uniformity in Un*x-based environments.  This is another one of those never-
ending silly issues that we might have to revisit again.