Prevent escaping of escape chars on output.
Change related to 807c36acefa1073c76b79208a1fbcbd595783bab
We are setting escapechar back to None to avoid escaping them on
output as in 'Buffalo \'66'.  The assumption is that all fields
are properly KGTK-quoted and escaped.  This also means any unquoted
TABs or NLs in fields will now raise errors as they should.  To
output them, they can be generated by explicitly creating their
escape sequences as in --return '"a\\tb"'.