Initial implementation of optional match.
Completely revamps how we translate joins for all match clauses,
strict and now also optional.
Runs the test suite fine which is encouraging, so at least we didn't break
anything obvious, but this requires much more testing to make sure we didn't
introduce any performance issues.
We now detect problematic queries that enumerate cross products, and there
is a new --force option to override that to run such queries anyway.
Improved bookkeeping for restrictions and joins on a per match-clause basis,
so we don't need to pass around that info as parameters everywhere.
Streamlined auto indexing and 'ensure_relevant_indexes' to not require
info that we can now access directly from the match clauses.
This also changed the API to 'translate_to_sql' which means we'll have to
update the browser interface accordingly when the time comes.