Initial support for querying over multiple main + auxiliary DBs
For now we only support read-only processing when auxiliary DBs are
attached for simplicity and to avoid having to adapt all DB-modifying
operations to support potentially qualified DB tables.  Down the road
we might want to relax that to support updates on the main DB even if
auxiliary DBs are attached.

For now things have been generalized on a minimal level to support
what's absolutely needed to query across DBs but not more.  It is
possible that we haven't found all places yet that need to be
appropriately generalized, but things seem to be working for now.

We also added a check to prevent generation of standard table indexes
on vector columns.