Fixed bug with translation of nested optionals
For certain complex optional clauses that required a join and
evaluation of a local where condition before the result could be left
joined to the prior match clause, variables in the where condition
might have been translated into a clause-external table reference
which caused SQL to break.  This is now fixed by looking for a
clause-local variable translation first.